requirements.txt
sentence-transformers
torch
tiktoken
markitdown
python-docx
pymupdf
numpy

